requirements.txt
google-generativeai>=0.7.2
python-dotenv>=1.0.1
pydub>=0.25.1
ffmpeg-python>=0.2.0
gTTS>=2.5.3
click>=8.1.7
rich>=13.7.1
python-slugify>=8.0.4
srt>=3.5.3
mutagen>=1.47.0
imageio-ffmpeg>=0.5.1
